allow printing tmdiff with %zu

+++ test/test_conversions_16.cpp
@@ -390,7 +390,7 @@ TestOneConversion16
         (uint16_t) tmdiff, src.size(), dst.size(), mups, mpps, reps, algoname);
     // grepable string with all the above infos, for import in to database.
     // algo function time srcsize dstsize testfile
-    printf("@utf_utils;%s;%s;%zu;%zu;%zu;%s\n","u8u16",algoname,tmdiff,src.size(),dst.size(),current_test_file.c_str());
+    printf("@utf_utils;%s;%s;%zu;%zu;%zu;%s\n","u8u16",algoname,(size_t)tmdiff,src.size(),dst.size(),current_test_file.c_str());
     if (dst != answer)
     {
         printf("error: result for %s differs from iconv()\n", name);
